airflow/requirements.txt
apache-airflow-providers-apache-spark==2.0.0
apache-airflow-providers-cncf-kubernetes==2.0.2